+++ module/strategy/WalkToBall/README.md
@@ -1,4 +1,4 @@
-# StrategiseWalk
+# WalkToBall
 
 ## Description
