+++ module/strategy/WalkToBall/src/WalkToBall.hpp
@@ -1,5 +1,5 @@
-#ifndef MODULE_STRATEGY_STRATEGISEWALK_HPP
-#define MODULE_STRATEGY_STRATEGISEWALK_HPP
+#ifndef MODULE_STRATEGY_WALKTOBALL_HPP
+#define MODULE_STRATEGY_WALKTOBALL_HPP
 
 #include <nuclear>
 
@@ -7,7 +7,7 @@
 
 namespace module::strategy {
 
-    class StrategiseWalk : public ::extension::behaviour::BehaviourReactor {
+    class WalkToBall : public ::extension::behaviour::BehaviourReactor {
     private:
         /// @brief Stores configuration values
         struct Config {
@@ -18,10 +18,10 @@ namespace module::strategy {
         } cfg;
 
     public:
-        /// @brief Called by the powerplant to build and setup the StrategiseWalk reactor.
-        explicit StrategiseWalk(std::unique_ptr<NUClear::Environment> environment);
+        /// @brief Called by the powerplant to build and setup the WalkToBall reactor.
+        explicit WalkToBall(std::unique_ptr<NUClear::Environment> environment);
     };
 
 }  // namespace module::strategy
 
-#endif  // MODULE_STRATEGY_STRATEGISEWALK_HPP
+#endif  // MODULE_STRATEGY_WALKTOBALL_HPP
