+++ roles/webots/directormotion.role
@@ -34,5 +34,5 @@ nuclear_role(
   # Strategies
   strategy::FallRecovery
   strategy::StrategiseLook
-  strategy::StrategiseWalk
+  strategy::WalkToBall
 )
